+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+- `genomepy install` no longer requires a provider argument. Providers may still be passed with the `-p/--provider` flag.
+- If no provider is passed to `genomepy install`, the first provider with the genome is used (order: Ensembl > UCSC > NCBI).
+
+### Added
+- automatic provider selection for `genomepy install`
+
 ## [0.8.4] - 2020-07-29
 - Fix bug where Genome.sizes dict contains str instead of int (#110).
 - Fix bug with UTF-8 in README (#109).