+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+- check to see if providers are online + error message if not
+
+### Changed
+- `genomepy.utils.check_url` now has a timeout and a retry system
+- `genomepy search` and `genomepy providers` now only check online providers
+
 ## [0.8.4] - 2020-07-29
 - Fix bug where Genome.sizes dict contains str instead of int (#110).
 - Fix bug with UTF-8 in README (#109).